+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 
 ### Changed
 - Add create_time and update_time to returned storage engine writes acks.
-- Add storage index create flag to read only from the index.
+- Add storageIndexList indexOnly param to read data only from the index.
 
 ### [3.17.1] - 2023-08-23
 ### Added